semester-5/scripting-langs/lab11/requirements.txt
Pillow==2.2.2
ttkwidgets